requirements.txt
alembic==0.9.6
authy==2.2.4
Babel==2.5.1
blinker==1.4
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
dominate==2.3.1
elasticsearch==7.1.0
Flask==1.1.1
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==3.3.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
guess-language-spirit==0.5.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10.3
Mako==1.0.7
MarkupSafe==1.0
PyJWT==1.7.1
python-dateutil==2.6.1
python-dotenv==0.10.3
python-editor==1.0.3
pytz==2017.2
qrcode==6.1
redis==3.2.1
requests==2.22.0
rq==1.1.0
simplejson==3.17.0
six==1.11.0
SQLAlchemy==1.1.14
urllib3==1.22
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.1
